Move getEntityWithName to model component
