+++ src/main/java/seedu/address/experimental/model/Model.java
@@ -86,6 +86,13 @@ public interface Model {
      */
     void setEntity(Entity target, Entity editedEntity);
 
+    /**
+     * Gets the entity with name matching the provided name.
+     * @param name the name of the entity to be found
+     * @return the found entity
+     */
+    Entity getEntityWithName(String name);
+
     // ============== Filtered entity list =================
     /**
      * Returns the entities listed by the given predicate
