+++ src/main/java/seedu/address/experimental/model/ModelManager.java
@@ -13,11 +13,8 @@ import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
+import seedu.address.model.entity.*;
 import seedu.address.model.entity.Character;
-import seedu.address.model.entity.Entity;
-import seedu.address.model.entity.Item;
-import seedu.address.model.entity.Mob;
-import seedu.address.model.entity.Template;
 import seedu.address.model.util.ComposedPredicate;
 
 /**
@@ -104,6 +101,11 @@ public class ModelManager implements Model {
         userPrefs.setRerollFilePath(rerollFilePath);
     }
 
+    @Override
+    public Predicate<Entity> getClassificationPredicate(Classification classification) {
+        return null;
+    }
+
     //=====================================Rerollllll==============================
 
     @Override
@@ -127,6 +129,20 @@ public class ModelManager implements Model {
         requireNonNull(entity);
         reroll.addEntity(entity);
         updateFilteredEntityList(PREDICATE_SHOW_ALL_ENTITIES);
+    }
+
+    @Override
+    public Entity getEntityWithName(String name) {
+        requireNonNull(name);
+        Entity toReturn = null;
+        ObservableList<Entity> entities = reroll.getEntities().getEntityList();
+        for (Entity entity : entities) {
+            if (entity.getName().fullName.equals(name)) {
+                toReturn = entity;
+                break;
+            }
+        }
+        return toReturn;
     }
 
     @Override
