+++ src/main/java/seedu/address/experimental/model/ReadOnlyEntities.java
@@ -13,6 +13,4 @@ public interface ReadOnlyEntities {
      * The list will not contain duplicate items.
      */
     ObservableList<Entity> getEntityList();
-
-    Entity getEntityWithName(String name);
 }
